# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: github.com/is7qin/c3api Admin API
  version: 1.0.0
  description: '管理端 API（模板/账号/分组/规则/日志/统计）。认证：Authorization: Bearer <admin_token>。'
servers:
  - url: /admin
...
paths:
  /templates:
    post:
      operationId: PostTemplates
      tags: [templates]
      summary: 创建模板
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TemplateCreate' }
      responses:
        '200':
          description: 创建后的模板
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
        default:
          $ref: '#/components/responses/Error'
    get:
      operationId: GetTemplates
      tags: [templates]
      summary: 模板列表（分页/筛选/排序）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 200, description: 每页条数（上限 200，超限裁剪到 200） } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: name, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: 模板列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TemplateListResponse' }
        default:
          $ref: '#/components/responses/Error'
  /templates/batch-delete:
    post:
      operationId: PostTemplatesBatchDelete
      tags: [templates]
      summary: 批量删除模板（事务，全成或全败）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchDeleteBody' }
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchDeleteResponse' }
        default:
          $ref: '#/components/responses/Error'
  /templates/batch-update:
    post:
      operationId: PostTemplatesBatchUpdate
      tags: [templates]
      summary: 批量更新模板（fields 为任意字段子集）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BatchUpdateTemplatesBody' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchUpdateResponse' }
        default:
          $ref: '#/components/responses/Error'
  /templates/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetTemplatesId
      tags: [templates]
      responses:
        '200':
          description: 模板
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutTemplatesId
      tags: [templates]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TemplateCreate' }
      responses:
        '200':
          description: 更新后模板
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeleteTemplatesId
      tags: [templates]
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
...
  /user/keys:
    get:
      operationId: GetUserKeys
      tags: [user]
      summary: 我的 key 列表（分页/排序；key 明文长期可查看/复制）
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 200, description: 每页条数（上限 200，超限裁剪到 200） } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: name, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: key 列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KeyListResponse' }
        default:
          $ref: '#/components/responses/Error'
    post:
      operationId: PostUserKeys
      tags: [user]
      summary: 创建 key（组可选性校验：public 或已授予 private；明文长期回显）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KeyCreate' }
      responses:
        '200':
          description: 创建后的 key（含明文）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Key' }
        default:
          $ref: '#/components/responses/Error'
  /user/keys/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: GetUserKeysId
      tags: [user]
      summary: key 详情（仅本人；他人 key → 404）
      responses:
        '200':
          description: key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Key' }
        default:
          $ref: '#/components/responses/Error'
    put:
      operationId: PutUserKeysId
      tags: [user]
      summary: 更新 key（name/status/max_concurrency/quota；仅本人）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KeyUpdate' }
      responses:
        '200':
          description: 更新后的 key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Key' }
        default:
          $ref: '#/components/responses/Error'
    delete:
      operationId: DeleteUserKeysId
      tags: [user]
      summary: 删除 key（仅本人；Auth 快照增量移除——立即失效）
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletedResponse' }
        default:
          $ref: '#/components/responses/Error'
  /user/keys/{id}/rotate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      operationId: PostUserKeysIdRotate
      tags: [user]
      summary: 轮换 key（仅本人；新明文生效，旧 key 立即失效）
      responses:
        '200':
          description: 轮换后的 key（含新明文）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Key' }
        default:
          $ref: '#/components/responses/Error'
...
components:
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    DeletedResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: boolean }
...
    TemplateCreate:
      type: object
      required: [name, supported_formats]
      properties:
        name: { type: string }
        # base_url 全类型可选（2026-08-14 用户裁决）；提供时须为裸根（不含 /v1，
        # service 校验）。空：codex 类型走 SDK 默认端点；api_key 静态透传留空
        # 则路由时失败（管理面不拦截）。
        base_url: { type: string }
        # credential_type 模板级：一个模板 = 一种号池；缺省 = api_key。
        # 生态三类型（responses-special/codex-oauth/codex-pat）只支持 resp / resp-ws /
        # openai-images / openai-search 格式（service 校验）；类型化配置走 /templates/{id}/ext。
        credential_type: { type: string, enum: [api_key, responses-special, codex-oauth, codex-pat], default: api_key }
        supported_formats:
          type: array
          minItems: 1
          items: { type: string, enum: [openai-chat, openai-responses, openai-responses-ws, openai-images, openai-search, anthropic] }
        models:
          type: array
          items: { type: string }
        format_models:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        model_mapping:
          type: object
          additionalProperties: { type: string }
    Template:
      type: object
      required: [ID, Name, BaseURL, SupportedFormats, CreatedAt, UpdatedAt]
      properties:
        ID: { type: integer, format: int64 }
        Name: { type: string }
        BaseURL: { type: string }
        CredentialType: { type: string, enum: [api_key, responses-special, codex-oauth, codex-pat], description: 模板号池类型；生态三类型只支持 resp / resp-ws 格式 }
        SupportedFormats:
          type: array
          items: { type: string, enum: [openai-chat, openai-responses, openai-responses-ws, openai-images, openai-search, anthropic] }
        Models:
          type: array
          items: { type: string }
        FormatModels:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        ModelMapping:
          type: object
          additionalProperties: { type: string }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表过滤已删；GET 单个可查已删项） }
...
    TemplateListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Template' }
...
    KeyStatus:
      type: string
      enum: [active, disabled]
    Key:
      type: object
      properties:
        ID: { type: integer, format: int64 }
        UserID: { type: integer, format: int64 }
        GroupID: { type: integer, format: int64 }
        Name: { type: string }
        key: { type: string, description: key 明文（长期可查看/复制；DB 泄露即暴露——自托管权衡，用户裁决） }
        Status: { $ref: '#/components/schemas/KeyStatus' }
        MaxConcurrency: { type: integer }
        Quota: { type: integer, format: int64, description: 累计 token 上限；0 = 不限 }
        QuotaUsed: { type: integer, format: int64, description: 已消耗（后扣；无额度 key 恒 0） }
        CreatedAt: { type: string, format: date-time }
        UpdatedAt: { type: string, format: date-time }
        DeletedAt: { type: string, format: date-time, nullable: true, description: 软删除时间戳；null = 存活（列表/鉴权过滤已删；GET 单个可查已删项） }
    KeyCreate:
      type: object
      required: [name, group_id]
      properties:
        name: { type: string }
        group_id: { type: integer, format: int64 }
        max_concurrency: { type: integer, description: 单 key 在途上限；0 = 不限 }
        quota: { type: integer, format: int64, description: 累计 token 上限；0 = 不限 }
    KeyUpdate:
      type: object
      properties:
        name: { type: string }
        status: { $ref: '#/components/schemas/KeyStatus' }
        max_concurrency: { type: integer }
        quota: { type: integer, format: int64 }
    KeyListResponse:
      type: object
      required: [total, rows]
      properties:
        total: { type: integer, format: int64 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/Key' }
...
    BatchDeleteBody:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
    BatchUpdateTemplatesBody:
      type: object
      required: [ids, fields]
      properties:
        ids:
          type: array
          items: { type: integer, format: int64 }
          minItems: 1
          maxItems: 100
        fields:
          $ref: '#/components/schemas/TemplatePatch'
    TemplatePatch:
      type: object
      properties:
        name: { type: string }
        base_url: { type: string }
        supported_formats:
          type: array
          minItems: 1
          items: { type: string, enum: [openai-chat, openai-responses, openai-responses-ws, openai-images, openai-search, anthropic] }
        models:
          type: array
          items: { type: string }
        format_models:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        model_mapping:
          type: object
          additionalProperties: { type: string }
...
    BatchDeleteResponse:
      type: object
      required: [deleted]
      properties:
        deleted: { type: integer }
    BatchUpdateResponse:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
    # --- 兑换码（Phase 5 计费前基础设施：资源发放的通用载体） ---
...